:package: change dto package structure, modify variable name [#64]
