+++ src/main/java/com/nailshop/nailborhood/controller/shop/ShopController.java
@@ -4,7 +4,7 @@ import com.nailshop.nailborhood.dto.artboard.response.ShopArtBoardListLookupResp
 import com.nailshop.nailborhood.dto.common.CommonResponseDto;
 import com.nailshop.nailborhood.dto.common.ResultDto;
 import com.nailshop.nailborhood.dto.review.response.ShopReviewListLookupResponseDto;
-import com.nailshop.nailborhood.dto.shop.response.ShopDetailListResponseDto;
+import com.nailshop.nailborhood.dto.shop.response.detail.ShopDetailListResponseDto;
 import com.nailshop.nailborhood.dto.shop.response.ShopListResponseDto;
 import com.nailshop.nailborhood.service.shop.ShopArtBoardListService;
 import com.nailshop.nailborhood.service.shop.ShopDetailService;
@@ -51,7 +51,7 @@ public class ShopController {
                                                                             @RequestParam(value = "size", defaultValue = "10", required = false) int size,
                                                                             @RequestParam(value = "orderby", defaultValue = "createdAt", required = false) String criteria,
                                                                             @RequestParam(value = "sort", defaultValue = "DESC", required = false) String sort) {
-        CommonResponseDto<Object> allShopsList = shopListLookupLocalService.getShopListbyDong( page, size, sort, criteria, dongId);
+        CommonResponseDto<Object> allShopsList = shopListLookupLocalService.getShopListByDong( page, size, sort, criteria, dongId);
         ResultDto<ShopListResponseDto> resultDto = ResultDto.in(allShopsList.getStatus(), allShopsList.getMessage());
         resultDto.setData((ShopListResponseDto) allShopsList.getData());
