+++ src/main/java/com/nailshop/nailborhood/dto/shop/response/detail/MenuDetailResponseDto.java
@@ -1,17 +1,17 @@
-package com.nailshop.nailborhood.dto.shop.response;
+package com.nailshop.nailborhood.dto.shop.response.detail;
 
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@Builder
 @NoArgsConstructor
 public class MenuDetailResponseDto {
     private Long menuId;
     private String name;
     private String price;
 
+    @Builder
     public MenuDetailResponseDto(Long menuId, String name, String price) {
         this.menuId = menuId;
         this.name = name;
