+++ src/main/java/com/nailshop/nailborhood/dto/shop/response/detail/ShopImgListResponseDto.java
@@ -1,4 +1,4 @@
-package com.nailshop.nailborhood.dto.shop.response;
+package com.nailshop.nailborhood.dto.shop.response.detail;
 
 import jakarta.persistence.Column;
 import lombok.Builder;
@@ -6,13 +6,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@Builder
 @NoArgsConstructor
 public class ShopImgListResponseDto {
 
     private String imgPath;
     private int imgNum;
 
+    @Builder
     public ShopImgListResponseDto(String imgPath, int imgNum) {
 
         this.imgPath = imgPath;
