+++ src/main/java/com/nailshop/nailborhood/service/shop/admin/AllShopsLookupAdminService.java
@@ -3,13 +3,12 @@ package com.nailshop.nailborhood.service.shop.admin;
 import com.nailshop.nailborhood.domain.shop.Shop;
 import com.nailshop.nailborhood.dto.common.CommonResponseDto;
 import com.nailshop.nailborhood.dto.common.PaginationDto;
-import com.nailshop.nailborhood.dto.shop.response.AllShopsListResponseDto;
-import com.nailshop.nailborhood.dto.shop.response.AllShopsLookupResponseDto;
+import com.nailshop.nailborhood.dto.shop.response.admin.AllShopsListResponseDto;
+import com.nailshop.nailborhood.dto.shop.response.admin.AllShopsLookupResponseDto;
 import com.nailshop.nailborhood.exception.NotFoundException;
 import com.nailshop.nailborhood.repository.review.ReviewRepository;
 import com.nailshop.nailborhood.repository.shop.DongRepository;
 import com.nailshop.nailborhood.repository.shop.MenuRepository;
-import com.nailshop.nailborhood.repository.shop.ShopImgRepository;
 import com.nailshop.nailborhood.repository.shop.ShopRepository;
 import com.nailshop.nailborhood.service.common.CommonService;
 import com.nailshop.nailborhood.type.ErrorCode;
