+++ src/test/java/com/nailshop/nailborhood/service/shop/admin/AllShopsLookupAdminServiceTest.java
@@ -3,8 +3,8 @@ package com.nailshop.nailborhood.service.shop.admin;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.nailshop.nailborhood.dto.common.ResultDto;
-import com.nailshop.nailborhood.dto.shop.response.AllShopsListResponseDto;
-import com.nailshop.nailborhood.dto.shop.response.AllShopsLookupResponseDto;
+import com.nailshop.nailborhood.dto.shop.response.admin.AllShopsListResponseDto;
+import com.nailshop.nailborhood.dto.shop.response.admin.AllShopsLookupResponseDto;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
